two more tasks

+++ src/test/java/com/apismensky/reversepolish/OperationTest.java
@@ -10,28 +10,28 @@ public class OperationTest {
     public void testMinus() {
         Operation operation = Operation.value("-");
         assertEquals(Operation.MINUS, operation);
-        assertEquals(-1, operation.getAction().apply(4, 5));
+        assertEquals(-1, (int) operation.getAction().apply(4, 5));
     }
 
     @Test
     public void testPlus() {
         Operation operation = Operation.value("+");
         assertEquals(Operation.PLUS, operation);
-        assertEquals(2, operation.getAction().apply(1, 1));
+        assertEquals(2, (int) operation.getAction().apply(1, 1));
     }
 
     @Test
     public void testDiv() {
         Operation operation = Operation.value("/");
         assertEquals(Operation.DIV, operation);
-        assertEquals(3, operation.getAction().apply(15, 5));
+        assertEquals(3, (int) operation.getAction().apply(15, 5));
     }
 
     @Test
     public void testMult() {
         Operation operation = Operation.value("*");
         assertEquals(Operation.MULT, operation);
-        assertEquals(35, operation.getAction().apply(5, 7));
+        assertEquals(35, (int) operation.getAction().apply(5, 7));
     }
 
     @Test
